% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleGUI.R
\name{exampleGUI}
\alias{exampleGUI}
\title{A Shiny-based GUI for the rs_IDs function}
\usage{
exampleGUI()
}
\description{
Launches a (possibly buggy) Shiny app that acts as a graphical user interface for the \code{\link{rs_IDs}} function. It's a bit hacky, so its performance is not guaranteed.
}
\examples{
\dontrun{
exampleGUI()  #yeah, just run it.
}
}
\seealso{
\code{\link{rs_IDs}} and the vignette \href{../doc/loosebirdtag.html}{\code{loosebirdtag}}.
}
\author{
Andrew Burchill, \email{andrew.burchill@asu.edu}
}
